+++ backend/mora/handler/impl/manager.py
@@ -5,6 +5,7 @@ from structlog import get_logger
 from asyncio import create_task, gather
 from datetime import datetime
 from typing import Any, Awaitable, Dict, Iterable, Optional
+from functools import partial
 
 from .. import reading
 from ... import mapping
@@ -57,64 +58,64 @@ class ManagerReader(reading.OrgFunkReadingHandler):
 
     @classmethod
     async def _get_mo_object_from_effect(cls, effect, start, end, funcid):
-
         person = mapping.USER_FIELD.get_uuid(effect)
         manager_type = mapping.ORG_FUNK_TYPE_FIELD.get_uuid(effect)
         manager_level = mapping.MANAGER_LEVEL_FIELD.get_uuid(effect)
         responsibilities = list(mapping.RESPONSIBILITY_FIELD.get_uuids(effect))
         org_unit = mapping.ASSOCIATED_ORG_UNIT_FIELD.get_uuid(effect)
-
-        base_obj = create_task(
-            super()._get_mo_object_from_effect(effect, start, end, funcid))
         only_primary_uuid = util.get_args_flag('only_primary_uuid')
 
+        async def return_none():
+            return None
+
+        facet_get_one_class_full = partial(
+            facet.request_bulked_get_one_class_full,
+            only_primary_uuid=only_primary_uuid
+        )
+
+        base_obj = create_task(
+            super()._get_mo_object_from_effect(effect, start, end, funcid)
+        )
+
+        person_task = return_none()
         if person:
             person_task = create_task(
                 employee.request_bulked_get_one_employee(
                     person,
-                    only_primary_uuid=only_primary_uuid))
+                    only_primary_uuid=only_primary_uuid
+                )
+            )
 
+        manager_type_task = return_none()
         if manager_type:
             manager_type_task = create_task(
-                facet.request_bulked_get_one_class_full(
-                    manager_type,
-                    only_primary_uuid=only_primary_uuid))
+                facet_get_one_class_full(manager_type)
+            )
 
+        manager_level_task = return_none()
         if manager_level:
             manager_level_task = create_task(
-                facet.request_bulked_get_one_class_full(
-                    manager_level,
-                    only_primary_uuid=only_primary_uuid))
+                facet_get_one_class_full(manager_level)
+            )
 
-        resp_tasks: Iterable[Awaitable] = await gather(
-            *[create_task(facet.request_bulked_get_one_class_full(
-                obj_uuid, only_primary_uuid=only_primary_uuid))
-                for obj_uuid in responsibilities])
-
-        org_unit_task = create_task(orgunit.request_bulked_get_one_orgunit(
-            org_unit, details=orgunit.UnitDetails.MINIMAL,
-            only_primary_uuid=only_primary_uuid
+        resp_tasks: Iterable[Awaitable] = map(create_task, map(
+            facet_get_one_class_full, responsibilities
         ))
+
+        org_unit_task = create_task(
+            orgunit.request_bulked_get_one_orgunit(
+                org_unit,
+                details=orgunit.UnitDetails.MINIMAL,
+                only_primary_uuid=only_primary_uuid
+            )
+        )
 
         func: Dict[Any, Any] = {
             **await base_obj,
-            mapping.RESPONSIBILITY: gather(*resp_tasks),
+            mapping.RESPONSIBILITY: gather(*await gather(*resp_tasks)),
             mapping.ORG_UNIT: await org_unit_task,
+            mapping.PERSON: await person_task,
+            mapping.MANAGER_TYPE: await manager_type_task,
+            mapping.MANAGER_LEVEL: await manager_level_task,
         }
-
-        if person:
-            func[mapping.PERSON] = await person_task
-        else:
-            func[mapping.PERSON] = None
-
-        if manager_type:
-            func[mapping.MANAGER_TYPE] = await manager_type_task
-        else:
-            func[mapping.MANAGER_TYPE] = None
-
-        if manager_level:
-            func[mapping.MANAGER_LEVEL] = await manager_level_task
-        else:
-            func[mapping.MANAGER_LEVEL] = None
-
         return func