+++ backend/mora/service/address_handler/dar.py
@@ -5,6 +5,7 @@ import requests
 import uuid
 
 from structlog import get_logger
+from aiohttp import ClientSession
 
 from . import base
 from ..validation.validator import forceable
@@ -25,7 +26,7 @@ class DARAddressHandler(base.AddressHandler):
     prefix = 'urn:dar:'
 
     @classmethod
-    def from_effect(cls, effect):
+    async def afrom_effect(cls, effect):
         """
         Initialize handler from LoRa object
 
@@ -33,25 +34,25 @@ class DARAddressHandler(base.AddressHandler):
         gracefully and return _some_ kind of result
         """
         # Cut off the prefix
-        handler = super().from_effect(effect)
+        handler = await super().afrom_effect(effect)
+
+        handler._name = NOT_FOUND
+        handler._href = None
 
         try:
-            address_object = handler._fetch_from_dar(handler.value)
+            address_object = await handler._afetch_from_dar(handler.value)
             handler._name = ''.join(
-                handler._address_string_chunks(address_object))
-            handler._href = (
-                'https://www.openstreetmap.org/'
-                '?mlon={x}&mlat={y}&zoom=16'.format(**address_object)
-                if 'x' in address_object and 'y' in address_object
-                else None
+                handler._address_string_chunks(address_object)
             )
+            if 'x' in address_object and 'y' in address_object:
+                handler._href = (
+                    'https://www.openstreetmap.org/'
+                    '?mlon={x}&mlat={y}&zoom=16'.format(**address_object)
+                )
         except LookupError:
             logger.warning(
                 'address lookup failed', handler_value=handler.value
             )
-
-            handler._name = NOT_FOUND
-            handler._href = None
 
         return handler
 
@@ -78,36 +79,34 @@ class DARAddressHandler(base.AddressHandler):
         return self._href
 
     @staticmethod
-    def _fetch_from_dar(addrid):
-        for addrtype in (
-            'adresser', 'adgangsadresser',
-            'historik/adresser', 'historik/adgangsadresser'
-        ):
-            try:
-                r = session.get(
-                    'https://dawa.aws.dk/' + addrtype,
-                    # use a list to work around unordered dicts in Python < 3.6
-                    params=[
-                        ('id', addrid),
-                        ('noformat', '1'),
-                        ('struktur', 'mini'),
-                    ],
-                )
+    async def _afetch_from_dar(addrid):
+        async with ClientSession() as session:
+            for addrtype in (
+                'adresser', 'adgangsadresser',
+                'historik/adresser', 'historik/adgangsadresser'
+            ):
+                try:
+                    response = await session.get(
+                        'https://dawa.aws.dk/' + addrtype,
+                        # use a list to work around unordered dicts in Python < 3.6
+                        params=[
+                            ('id', addrid),
+                            ('noformat', '1'),
+                            ('struktur', 'mini'),
+                        ],
+                    )
+                    response.raise_for_status()
+                    addrobjs = await response.json()
 
-                addrobjs = r.json()
-
-                r.raise_for_status()
-
-                if addrobjs:
-                    # found, escape loop!
-                    break
-            # The request mocking library throws a pretty generic exception
-            # catch and rethrow as something we know how to manage
-            except Exception as e:
-                raise LookupError(str(e)) from e
-
-        else:
-            raise LookupError('no such address {!r}'.format(addrid))
+                    if addrobjs:
+                        # found, escape loop!
+                        break
+                # The request mocking library throws a pretty generic exception
+                # catch and rethrow as something we know how to manage
+                except Exception as e:
+                    raise LookupError(str(e)) from e
+            else:
+                raise LookupError('no such address {!r}'.format(addrid))
 
         return addrobjs.pop()